move StringDesc to moko-resources
