+++ mvvm/src/androidMain/kotlin/dev/icerock/moko/mvvm/databinding/ViewBindingAdapters.kt
@@ -2,7 +2,7 @@
  * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.mvvm.desc.databinding
+package dev.icerock.moko.mvvm.databinding
 
 import android.view.View
 import androidx.databinding.BindingAdapter
